refactor: abstract delete buttons
